Run the release workflow on tag
The previous release workflow was manually triggered, and would create the
release manifests. That's nice in theory, but there's no good way to let Actions
push to a protected branch.

Run the release manifest when a tag is created instead, and have it validate
that the release manifests exist before pushing.